Actualizacion del programa
Alta de extraordinarios realizado

+++ src/main/java/unam/fes/model/entity/SemestreEntity.java
@@ -9,18 +9,18 @@ public class SemestreEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_semestre")
-    private Integer idSemestre;
+    private Long idSemestre;
 
     @Column(name = "semestre", length = 8, nullable = false)
     private String semestre;
 
     // Getters y setters
 
-    public Integer getIdSemestre() {
+    public Long getIdSemestre() {
         return idSemestre;
     }
 
-    public void setIdSemestre(Integer idSemestre) {
+    public void setIdSemestre(Long idSemestre) {
         this.idSemestre = idSemestre;
     }
